refactor QuestionsAnwers.jsx to use hooks, cleaned up questionsList directory
