+++ client/src/components/questions-answers/questionsList/QuestionListItem.jsx
@@ -3,18 +3,10 @@ import AnswersPerQuestion from './AnswersPerQuestion';
 
 const QuestionListItem = props => {
 
-  // console.log('ANSWERS')
-  // // console.log(props.question.answers)
   let answers = props.question.answers;
-
-  // Object.keys(answers).map((key) => {
-  //   console.log(answers[key].body)
-  // })
-
 
   return (
     <>
-
       <div>
         <li className="QuestionItem">
           Q: {props.question.question_body}
